move: db-migrator into gcc
